Move logic from infrastructure to domain package
Now in domain croupier is also a player like it was saved in database.
Moved methods which return croupier and current player id to domain from
infrastructure. Added toString method to PlayerGameId for better readability
of exception message.
Commented part in GameEntity constructor which was responsible for saving
players in database - currently it can not work, because for saving players
which are added to entity game must be already created in db.