Remove business logic from GameController
WIth this change GameController is not responsible for
checking that player exist in database, and DO NOT provide
croupier when creating game.
Also with this change renamed all forms, to command classes.
Now GameController is responsible for creating only GameService
command objects and execute GameService methods.
The effect of these changes is that GameService now has dependency
to PlayerService and is responsible for calling this service when
need get player from database.